fixed directory creation before saving output file
