+++ src/every_query/aces_to_eq/aces_to_eq.py
@@ -1,7 +1,5 @@
-import hashlib
 import logging
 import os
-import re
 from pathlib import Path
 
 import hydra
@@ -9,14 +7,10 @@ import polars as pl
 from omegaconf import DictConfig
 from tqdm import tqdm
 
+from every_query.utils.codes import code_slug
+
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO)
-
-
-def code_slug(code: str, n_hash: int = 10, prefix_len: int = 24) -> str:
-    h = hashlib.sha1(code.encode("utf-8")).hexdigest()[:n_hash]
-    prefix = re.sub(r"[^A-Za-z0-9._-]+", "_", code).strip("_")[:prefix_len]
-    return f"{prefix}__{h}" if prefix else h
 
 
 def shard_id(p: Path) -> int:
