+++ src/every_query/eval_suite/gen_task.py
@@ -1,5 +1,3 @@
-import hashlib
-import re
 from pathlib import Path
 
 import hydra
@@ -7,11 +5,7 @@ import polars as pl
 from omegaconf import DictConfig
 from tqdm import tqdm
 
-
-def code_slug(code: str, n_hash: int = 10, prefix_len: int = 24) -> str:
-    h = hashlib.sha1(code.encode("utf-8")).hexdigest()[:n_hash]
-    prefix = re.sub(r"[^A-Za-z0-9._-]+", "_", code).strip("_")[:prefix_len]
-    return f"{prefix}__{h}" if prefix else h
+from every_query.utils.codes import code_slug
 
 
 def list_parquets(d: Path) -> list[Path]:
