+++ src/every_query/process_composite/get_per_code_from_composite.py
@@ -1,8 +1,5 @@
-import hashlib
 import logging
-import re
 from pathlib import Path
-from typing import Any
 
 import hydra
 import polars as pl
@@ -10,18 +7,10 @@ from omegaconf import DictConfig
 from sklearn.metrics import roc_auc_score
 from tqdm import tqdm
 
+from every_query.utils.codes import code_slug
+
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO)
-
-
-def values_as_list(**kwargs) -> list[Any]:
-    return list(kwargs.values())
-
-
-def code_slug(code: str, n_hash: int = 10, prefix_len: int = 24) -> str:
-    h = hashlib.sha1(code.encode("utf-8")).hexdigest()[:n_hash]
-    prefix = re.sub(r"[^A-Za-z0-9._-]+", "_", code).strip("_")[:prefix_len]
-    return f"{prefix}__{h}" if prefix else h
 
 
 @hydra.main(
